+++ changelog.rst
@@ -6,6 +6,8 @@ Release 1.0.21
 
 - Enhancement: Added :code:`rmr.utils.cache.CacheTimeout`, lazy value for :code:`django.views.decorators.cache.cache_page` decorator
 - Enhancement: Added custom cache backend :code:`rmr.utils.cache.RmrLibMCCache` which enables to use :code:`rmr.utils.cache.CacheTimeout` decorator
+- Enhancement: Added HTTP-caching headers management to the :code:`rmr.views.Json`
+- Enhancement: Moved :code:`validate_request` decorator to :code:`rmr.views.decorators.validation` module
 
 Release 1.0.20
 --------------