+++ changelog.rst
@@ -4,8 +4,7 @@ Changelog
 Release 1.0.21
 --------------
 
-- Enhancement: Added :code:`rmr.utils.cache.CacheTimeout`, lazy value for :code:`django.views.decorators.cache.cache_page` decorator
-- Enhancement: Added custom cache backend :code:`rmr.utils.cache.RmrLibMCCache` which enables to use :code:`rmr.utils.cache.CacheTimeout` decorator
+- Enhancement: Added 'Content-Length' setting on :code:`rmr.views.Json` responses
 - Enhancement: Added HTTP-caching headers management to the :code:`rmr.views.Json`
 - Enhancement: Moved :code:`validate_request` decorator to :code:`rmr.views.decorators.validation` module
 - Enhancement: Removed :code:`rmr.views.Json.get_device_id()` method